% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmde.r
\name{scimp_bmde}
\alias{scimp_bmde}
\title{Bayesian Multinomial Dirichlet Model (Equal Prior)}
\usage{
scimp_bmde(x, p, seed = 1492)
}
\arguments{
\item{x}{cell counts of given contingency table corresponding to a categorical data - non negative integers}

\item{p}{equal value for the Dirichlet prior parameter - positive real number}

\item{seed}{random seed for reproducible results}
}
\value{
\code{tibble} with original limits of multinomial proportions together with product of length of k intervals as volume of simultaneous confidence intervals and the mean
}
\description{
This method provides 95 percent simultaneous confidence interval for multinomial proportions based on Bayesian Multinomial Dirichlet model. However, it provides a mechanism through which user can split the Dirichlet prior parameter vector and suitable distributions can be incorporated for each of two groups.
}
\examples{
y <- c(44, 55, 43, 32, 67, 78)
z <- 1
scimp_bmde(y, z)
}
\references{
Gelman, A., Carlin, J.B., Stern, H.S., and Rubin, D.B. (2002). Bayesian Data Analysis. Chapman & Hall, London.
}
\author{
Dr M Subbiah
}
