% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdu.r
\name{scimp_bmdu}
\alias{scimp_bmdu}
\title{Bayesian Multinomial Dirichlet Model (Unequal Prior)}
\usage{
scimp_bmdu(x, d, seed = 1492)
}
\arguments{
\item{x}{cell counts of given contingency table corresponding to a categorical data - non negative integers}

\item{d}{number of divisions required to split the prior vector of Dirichlet distribution to assign unequal values from U(0,1) and U(1,2)}

\item{seed}{random seed for reproducible results}
}
\value{
\code{tibble} with original limits of multinomial proportions together with product of length of k intervals as volume of simultaneous confidence intervals and the mean
}
\description{
This method provides 95 percent simultaneous confidence interval for multinomial proportions based on Bayesian Multinomial Dirichlet model. However, it provides a mechanism through which user can split the Dirichlet prior parameter vector and suitable distributions can be incorporated for each of two groups.
}
\examples{
y <- c(44, 55, 43, 32, 67, 78)
z <- 2
scimp_bmdu(y, z)
}
\references{
Gelman, A., Carlin, J.B., Stern, H.S., and Rubin, D.B. (2002). Bayesian Data Analysis. Chapman & Hall, London.
}
\author{
Dr M Subbiah
}
