% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitz-scott.r
\name{scimp_fs}
\alias{scimp_fs}
\title{Fitzpatrick & Scott Confidence Interval}
\usage{
scimp_fs(inpmat, alpha)
}
\arguments{
\item{inpmat}{the cell counts of given contingency tables corresponding to categorical data}

\item{alpha}{a number in \verb{[0..1]} to get the upper 100(1-\code{alpha}) percentage point of the chi square distribution}
}
\value{
\code{tibble} with original and adjusted limits of multinomial proportions together with product of length of k intervals as volume of simultaneous confidence intervals
}
\description{
The simultaneous confidence interval for multinomial proportions based on the method proposed in Fitzpatrick and Scott (1987)
}
\examples{
y <- c(44, 55, 43, 32, 67, 78)
z <- 0.05
scimp_fs(y, z)
}
\references{
Fitzpatrick, S. and Scott, A. (1987). Quick simultaneous confidence interval for multinomial proportions. Journal of American Statistical Association 82(399): 875-878.
}
\author{
Dr M Subbiah
}
