% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodman.r
\name{scimp_goodman}
\alias{scimp_goodman}
\title{Goodman Confidence Interval}
\usage{
scimp_goodman(inpmat, alpha)
}
\arguments{
\item{inpmat}{the cell counts of given contingency tables corresponding to categorical data}

\item{alpha}{a number in \verb{[0..1]} to get the upper 100(1-\code{alpha}) percentage point of the chi square distribution}
}
\value{
\code{tibble} with original and adjusted limits of multinomial proportions together with product of length of k intervals as volume of simultaneous confidence intervals
}
\description{
The simultaneous confidence interval for multinomial proportions based on the method proposed in Goodman (1965)
}
\examples{
y <- c(44, 55, 43, 32, 67, 78)
z <- 0.05
scimp_goodman(y, z)
}
\references{
Goodman, L.A. (1965). On Simultaneous Confidence Intervals for Multinomial Proportions. Technometrics 7: 247-254.
}
\author{
Dr M Subbiah
}
