% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qh.r
\name{scimp_qh}
\alias{scimp_qh}
\title{Quesenberry & Hurst Confidence Interval}
\usage{
scimp_qh(inpmat, alpha)
}
\arguments{
\item{inpmat}{the cell counts of given contingency tables corresponding to categorical data}

\item{alpha}{a number in \code{[0..1]} to get the upper 100(1-\code{alpha}) percentage point of the chi square distribution}
}
\value{
\code{tibble} with original and adjusted limits of multinomial proportions together with product of length of k intervals as volume of simultaneous confidence intervals
}
\description{
The simultaneous confidence interval for multinomial proportions based on the method proposed in Quesenberry and Hurst (1964)
}
\examples{
y <- c(44, 55, 43, 32, 67, 78)
z <- 0.05
scimp_qh(y, z)
}
\references{
Quesensberry, C.P. and Hurst, D.C. (1964). Large Sample Simultaneous Confidence Intervals for Multinational Proportions. Technometrics, 6: 191-195.
}
\author{
Dr M Subbiah
}
