% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sison-glaz.r
\name{scimp_sg}
\alias{scimp_sg}
\title{Sison & Glaz Confidence Interval}
\usage{
scimp_sg(inpmat, alpha)
}
\arguments{
\item{inpmat}{the cell counts of given contingency tables corresponding to categorical data}

\item{alpha}{a number in \verb{[0..1]} to get the upper 100(1-\code{alpha}) percentage point of the chi square distribution}
}
\value{
\code{tibble} with original and adjusted limits of multinomial proportions together with product of length of k intervals as volume of simultaneous confidence intervals
}
\description{
The simultaneous confidence interval for multinomial proportions based on the method proposed in Sison and Glaz (1995).
}
\examples{
y <- c(44, 55, 43, 32, 67, 78)
z <- 0.05
scimp_sg(y, z)
}
\references{
Sison, P.C. and Glaz J. (1995). Simultaneous Confidence Intervals and Sample Size Determination for Multinomial Proportions. Journal of the American Statistical Association 90: 366-369.
}
\author{
Dr M Subbiah
}
