% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.r
\name{scimp_wald}
\alias{scimp_wald}
\title{Wald Confidence Interval}
\usage{
scimp_wald(inpmat, alpha)
}
\arguments{
\item{inpmat}{the cell counts of given contingency tables corresponding to categorical data}

\item{alpha}{a number in \code{[0..1]} to get the upper 100(1-\code{alpha}) percentage point of the chi square distribution}
}
\value{
\code{tibble} with original and adjusted limits of multinomial proportions together with product of length of k intervals as volume of simultaneous confidence intervals
}
\description{
The simple Wald type interval for multinomial proportions which is symmetrical about the
sample proportions. In this method no continuity corrections are made to avoid zero width
intervals when the sample proportions are at extreme.
}
\examples{
y <- c(44, 55, 43, 32, 67, 78)
z <- 0.05
scimp_wald(y, z)
}
\references{
Wald, A Tests of statistical hypotheses concerning several parameters when the number of observations is large, Trans. Am. Math. Soc. 54 (1943) 426-482.
}
\author{
Dr M Subbiah
}
