% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldcc.r
\name{scimp_waldcc}
\alias{scimp_waldcc}
\title{Wald Confidence Interval (with continuity correction)}
\usage{
scimp_waldcc(inpmat, alpha)
}
\arguments{
\item{inpmat}{the cell counts of given contingency tables corresponding to categorical data}

\item{alpha}{a number in \verb{[0..1]} to get the upper 100(1-\code{alpha}) percentage point of the chi square distribution}
}
\value{
\code{tibble} with original and adjusted limits of multinomial proportions together with product of length of k intervals as volume of simultaneous confidence intervals
}
\description{
The simple Wald type interval with continuity corrections for multinomial proportions which is symmetrical about the sample proportions.
}
\examples{
y <- c(44, 55, 43, 32, 67, 78)
z <- 0.05
scimp_waldcc(y, z)
}
\author{
Dr M Subbiah
}
