% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{scimple}
\alias{scimple}
\title{Simultaneous Confidence Intervals for Multinomial Proportions}
\description{
Methods for obtaining simultaneous confidence intervals for multinomial proportions have
been proposed by many authors and the present study include a variety of widely
applicable procedures. Seven classical methods (Wilson, Quesenberry and Hurst, Goodman,
Wald with and without continuity correction, Fitzpatrick and Scott, Sison and Glaz)
and Bayesian Dirichlet models are included in the package. The advantage of MCMC pack
has been exploited to derive the Dirichlet posterior directly and this also helps in
handling the Dirichlet prior parameters. This package is prepared to have equal and
unequal values for the Dirichlet prior distribution that will provide better scope for
data analysis and associated sensitivity analysis.
}
\author{
Dr M.Subbiah [primary], Bob Rudis (bob@rud.is) [tidy version]
}
