% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scimple.r
\name{scimple_ci}
\alias{scimple_ci}
\title{Calculate multiple simultaneous confidence intervals using selected methods (excluding Bayesian methods)}
\usage{
scimple_ci(
  inpmat,
  alpha,
  methods = c("fs", "goodman", "qh", "sg", "wald", "waldcc", "wilson")
)
}
\arguments{
\item{inpmat}{the cell counts of given contingency tables corresponding to categorical data}

\item{alpha}{a number in \verb{[0..1]} to get the upper 100(1-\code{alpha}) percentage point of the chi square distribution}

\item{methods}{character vector of one or more methods to run over the input parameters}
}
\value{
\code{tibble} with original and adjusted limits of multinomial proportions together with
product of length of k intervals as volume of simultaneous confidence intervals.
The \code{inputmat} and \code{alpha} values are also returned in the \code{tibble}.
}
\description{
Return simultaneous confidence intervals for multinomial proportions based on selected methods.
}
\examples{
y <- c(44,55,43,32,67,78)
z <- 0.05
scimple_ci(y, z)
scimple_ci(y, z, c("goodman", "waldcc", "wilson"))
}
\author{
Bob Rudis (bob@rud.is)
}
