% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\docType{data}
\name{scimple_short_to_long}
\alias{scimple_short_to_long}
\title{Simple tranlsation table from method shorthand to full method name}
\format{
An object of class \code{character} of length 7.
}
\usage{
scimple_short_to_long
}
\description{
Useful for charts/reports
}
\examples{
scimple_short_to_long[c("waldcc", "sg", "qh")]
}
\keyword{datasets}
